Moved test utils to reaktive-test module
