+++ .gitignore
@@ -9,3 +9,4 @@
 */build
 */node_modules
 */package-lock.json
+*.hprof
