+++ reaktive-test/src/commonMain/kotlin/com/badoo/reaktive/test/observable/TestObservableExt.kt
@@ -1,3 +1,3 @@
-package com.badoo.reaktive.testutils
+package com.badoo.reaktive.test.observable
 
 val TestObservable<*>.hasSubscribers: Boolean get() = observers.isNotEmpty()
