+++ reaktive-test/src/commonTest/kotlin/com/badoo/reaktive/test/DummyTest.kt
@@ -1,0 +1,10 @@
+package com.badoo.reaktive.test
+
+import kotlin.test.Test
+
+class DummyTest {
+
+    @Test
+    fun dummy() {
+    }
+}
