+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/CombineLatestTest.kt
@@ -1,11 +1,11 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.dispose
-import com.badoo.reaktive.testutils.getOnNextValue
-import com.badoo.reaktive.testutils.hasOnNext
-import com.badoo.reaktive.testutils.isCompleted
-import com.badoo.reaktive.testutils.isError
-import com.badoo.reaktive.testutils.test
+import com.badoo.reaktive.test.observable.dispose
+import com.badoo.reaktive.test.observable.getOnNextValue
+import com.badoo.reaktive.test.observable.hasOnNext
+import com.badoo.reaktive.test.observable.isCompleted
+import com.badoo.reaktive.test.observable.isError
+import com.badoo.reaktive.test.observable.test
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
