+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/MapTest.kt
@@ -1,9 +1,9 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.TestObservable
-import com.badoo.reaktive.testutils.isError
-import com.badoo.reaktive.testutils.test
-import com.badoo.reaktive.testutils.values
+import com.badoo.reaktive.test.observable.TestObservable
+import com.badoo.reaktive.test.observable.isError
+import com.badoo.reaktive.test.observable.test
+import com.badoo.reaktive.test.observable.values
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
