+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/ObservableByEmitterTest.kt
@@ -1,16 +1,16 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.TestObservableObserver.Event
-import com.badoo.reaktive.testutils.dispose
-import com.badoo.reaktive.testutils.getOnErrorValue
-import com.badoo.reaktive.testutils.getOnNextEvent
-import com.badoo.reaktive.testutils.getOnNextValue
-import com.badoo.reaktive.testutils.hasOnNext
-import com.badoo.reaktive.testutils.isCompleted
-import com.badoo.reaktive.testutils.isDisposed
-import com.badoo.reaktive.testutils.isError
-import com.badoo.reaktive.testutils.isOnCompleteEvent
-import com.badoo.reaktive.testutils.test
+import com.badoo.reaktive.test.observable.TestObservableObserver.Event
+import com.badoo.reaktive.test.observable.dispose
+import com.badoo.reaktive.test.observable.getOnErrorValue
+import com.badoo.reaktive.test.observable.getOnNextEvent
+import com.badoo.reaktive.test.observable.getOnNextValue
+import com.badoo.reaktive.test.observable.hasOnNext
+import com.badoo.reaktive.test.observable.isCompleted
+import com.badoo.reaktive.test.observable.isDisposed
+import com.badoo.reaktive.test.observable.isError
+import com.badoo.reaktive.test.observable.isOnCompleteEvent
+import com.badoo.reaktive.test.observable.test
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
