+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/ObserveOnTest.kt
@@ -1,14 +1,14 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.TestObservable
-import com.badoo.reaktive.testutils.TestScheduler
-import com.badoo.reaktive.testutils.dispose
-import com.badoo.reaktive.testutils.hasOnNext
-import com.badoo.reaktive.testutils.hasSubscribers
-import com.badoo.reaktive.testutils.isCompleted
-import com.badoo.reaktive.testutils.isError
-import com.badoo.reaktive.testutils.test
-import com.badoo.reaktive.testutils.values
+import com.badoo.reaktive.test.observable.TestObservable
+import com.badoo.reaktive.test.scheduler.TestScheduler
+import com.badoo.reaktive.test.observable.dispose
+import com.badoo.reaktive.test.observable.hasOnNext
+import com.badoo.reaktive.test.observable.hasSubscribers
+import com.badoo.reaktive.test.observable.isCompleted
+import com.badoo.reaktive.test.observable.isError
+import com.badoo.reaktive.test.observable.test
+import com.badoo.reaktive.test.observable.values
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
