+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/SkipTest.kt
@@ -1,8 +1,8 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.TestObservable
-import com.badoo.reaktive.testutils.test
-import com.badoo.reaktive.testutils.values
+import com.badoo.reaktive.test.observable.TestObservable
+import com.badoo.reaktive.test.observable.test
+import com.badoo.reaktive.test.observable.values
 import kotlin.test.Test
 import kotlin.test.assertEquals
