+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/observable/UpstreamDownstreamGenericTests.kt
@@ -1,10 +1,10 @@
 package com.badoo.reaktive.observable
 
-import com.badoo.reaktive.testutils.TestObservable
-import com.badoo.reaktive.testutils.dispose
-import com.badoo.reaktive.testutils.isCompleted
-import com.badoo.reaktive.testutils.isError
-import com.badoo.reaktive.testutils.test
+import com.badoo.reaktive.test.observable.TestObservable
+import com.badoo.reaktive.test.observable.dispose
+import com.badoo.reaktive.test.observable.isCompleted
+import com.badoo.reaktive.test.observable.isError
+import com.badoo.reaktive.test.observable.test
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
